feat: 字体管理

+++ core/core-backend/src/main/java/io/dataease/font/manage/FontManage.java
@@ -6,8 +6,10 @@ import io.dataease.exception.DEException;
 import io.dataease.font.dao.auto.entity.CoreFont;
 import io.dataease.font.dao.auto.mapper.CoreFontMapper;
 import io.dataease.utils.BeanUtils;
+import io.dataease.utils.FileUtils;
 import io.dataease.utils.IDUtils;
 import jakarta.annotation.Resource;
+import org.apache.commons.lang3.ObjectUtils;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -41,22 +43,30 @@ public class FontManage {
         fontDto.setId(IDUtils.snowID());
         CoreFont coreFont = new CoreFont();
         BeanUtils.copyBean(coreFont, fontDto);
+        coreFont.setUpdateTime(System.currentTimeMillis());
         coreFontMapper.insert(coreFont);
         return fontDto;
     }
 
 
     public FontDto edit(FontDto fontDto) {
-        fontDto.setId(IDUtils.snowID());
+        if (ObjectUtils.isEmpty(fontDto.getId())) {
+            return create(fontDto);
+        }
         CoreFont coreFont = new CoreFont();
         BeanUtils.copyBean(coreFont, fontDto);
+        coreFont.setUpdateTime(System.currentTimeMillis());
         coreFontMapper.updateById(coreFont);
         return fontDto;
     }
 
     public void delete(Long id) {
-        coreFontMapper.deleteById(id);
-        //TODO delete file
+        CoreFont coreFont = coreFontMapper.selectById(id);
+        if (coreFont != null) {
+            coreFontMapper.deleteById(id);
+            FileUtils.deleteFile(path + coreFont.getFileTransName());
+        }
+
     }
 
     public void changeDefault(FontDto fontDto) {
@@ -67,16 +77,9 @@ public class FontManage {
         coreFontMapper.update(record, queryWrapper);
     }
 
-    public void upload(MultipartFile file, long fontID) {
-        String filename = file.getOriginalFilename();
-        QueryWrapper<CoreFont> queryWrapper = new QueryWrapper<>();
-        queryWrapper.eq("id", fontID);
-        CoreFont record = new CoreFont();
-        record.setFileName(filename);
-        record.setFileTransName(filename);
-        coreFontMapper.update(record, queryWrapper);
+    public String upload(MultipartFile file) {
         String fileUuid = UUID.randomUUID().toString();
-        saveFile(file, fileUuid);
+        return saveFile(file, fileUuid);
     }
 
     private static String saveFile(MultipartFile file, String fileNameUUID) throws DEException {
